Removed SimpleMultibandTile trait.
