--- raster/src/main/scala/geotrellis/raster/SimpleMultibandTile.scala
@@ -1,7 +1,0 @@
-package geotrellis.raster
-
-trait SimpleMultibandTile {
-  def bands(bandSequence: Seq[Int]): MultibandTile
-
-  def bands(bandSequence: Int*)(implicit d: DummyImplicit): MultibandTile
-}
